ci: cancel running CI workflows for current ref
This change uses the concurrency feature to cancel running CI for the ref the workflow is running on, avoiding stale CI runs (inspired by [this](https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#example-using-concurrency-to-cancel-any-in-progress-job-or-run)). I think using `github.ref` will make it work for any branch, whereas `github.head_ref` will only do it for PRs, because there's no `github.base_ref` for a branch without a PR. The main difference between these for repos that only run CI on PRs is whether stale CI workflows for `main` will be canceled.